requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
docopt==0.6.2
Flask==1.1.2
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==6.2.2
pytest-watch==4.2.0
python-docx==0.8.10
requests==2.25.1
toml==0.10.2
urllib3==1.26.3
watchdog==1.0.2
Werkzeug==1.0.1
